chore(6334): move lsp related types into module, with better division of purpose: client-inititated executeCommands, versus LspServer-initiated-commands
